+++ data-migrations/2025-07-30-0938-add-useSampleTextAsDefault-to-versionedQuestion.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `versionedQuestions`
+  ADD COLUMN `useSampleTextAsDefault` TINYINT(1) NOT NULL DEFAULT 0;