+++ src/resolver.ts
@@ -19,7 +19,6 @@ import { resolvers as planResolvers } from './resolvers/plan';
 import { resolvers as projectResolvers } from './resolvers/project';
 import { resolvers as questionResolvers } from './resolvers/question';
 import { resolvers as questionConditionResolvers } from './resolvers/questionCondition';
-import { resolvers as questionTypeResolvers } from './resolvers/questionType';
 import { resolvers as templateResolvers } from './resolvers/template';
 import { resolvers as repositoryResolvers } from './resolvers/repository';
 import { resolvers as researchDomainResolvers } from './resolvers/researchDomain';
@@ -55,7 +54,6 @@ export const resolvers: IResolvers = mergeResolvers([
   projectResolvers,
   questionResolvers,
   questionConditionResolvers,
-  questionTypeResolvers,
   repositoryResolvers,
   researchDomainResolvers,
   templateResolvers,